logout, login-updated, promote, demote, auth, validation, delete_user
