+++ backend/AAiT-backend-group-8/Delivery/Controller/blog_controller.go
@@ -2,7 +2,8 @@ package controller
 
 import (
 	"AAiT-backend-group-8/Domain"
-	"AAiT-backend-group-8/Helper"
+	infrastructure "AAiT-backend-group-8/Infrastructure"
+
 	"log"
 	"net/http"
 	"strconv"
@@ -26,7 +27,7 @@ func (controller *Controller) CreateBlog(ctx *gin.Context) {
 		return
 	}
 
-	claims, err := Helper.Parse(ctx)
+	claims, err := infrastructure.Parse(ctx)
 
 	if err != nil {
 		ctx.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
@@ -101,7 +102,7 @@ func (controller *Controller) UpdateBlog(ctx *gin.Context) {
 		return
 	}
 
-	claim, err := Helper.Parse(ctx)
+	claim, err := infrastructure.Parse(ctx)
 
 	if err != nil {
 		ctx.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
@@ -134,7 +135,7 @@ func (controller *Controller) DeleteBlog(ctx *gin.Context) {
 		return
 	}
 
-	claims, err := Helper.Parse(ctx)
+	claims, err := infrastructure.Parse(ctx)
 
 	if err != nil {
 		ctx.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
@@ -168,7 +169,7 @@ func (controller *Controller) DeleteBlog(ctx *gin.Context) {
 }
 
 func (controller *Controller) SearchBlog(ctx *gin.Context) {
-	searchParams := Helper.GetSearchParams(ctx)
+	searchParams := infrastructure.GetSearchParams(ctx)
 	blogs, err := controller.blogUseCase.SearchBlog(searchParams)
 
 	if err != nil {
