+++ backend/AAiT-backend-group-8/Domain/blog_domain.go
@@ -21,8 +21,8 @@ type Blog struct {
 }
 
 type SearchCriteria struct {
-	Title     string    `form:"title" bson:"title" json:"title" validate:"required, min=3,max=100"`
-	Author    string    `form:"author" bson:"author" json:"author" validate:"max=100"`
+	Title     string    `form:"title" bson:"title" json:"title"`
+	Author    string    `form:"author" bson:"author" json:"author"`
 	Tags      []string  `form:"tags" bson:"tags" json:"tags"`
 	StartDate time.Time `form:"startDate" bson:"start_date" json:"start_date"`
 	EndDate   time.Time `form:"endDate" bson:"end_date" json:"end_date"`
@@ -30,28 +30,4 @@ type SearchCriteria struct {
 	SortBy    string    `form:"sortBy" bson:"sort_by" json:"sort_by"`
 	Page      int       `form:"page" bson:"page" json:"page"`
 	PageSize  int       `form:"pageSize" bson:"page_size" json:"page_size"`
-}
-
-type IBlogRepository interface {
-	Search(criteria *SearchCriteria) ([]Blog, error)
-	Create(blog *Blog) error
-	UpdateCommentCount(id string, inc bool) error
-	UpdateLikeCount(id string, inc bool) error
-	FindAll(page int, pageSize int, sortBy string) ([]Blog, error)
-	FindByID(ID string) (*Blog, error)
-	Delete(ID string) error
-	UpdateViewCount(id string) error
-	Update(blog *Blog) error
-}
-
-type IBlogUseCase interface {
-	SearchBlog(criteria *SearchCriteria) ([]Blog, error)
-	CreateBlog(blog *Blog) error
-	UpdateBlogViewCount(id string) error
-	UpdateBlogCommentCount(id string, inc bool) error
-	UpdateBlogLikeCount(id string, inc bool) error
-	GetAllBlogs(page int, pageSize int, sortBy string) ([]Blog, error)
-	GetBlogByID(ID string) (*Blog, error)
-	DeleteBlog(ID string) error
-	UpdateBlog(blog *Blog) error
 }
