+++ backend/AAiT-backend-group-8/Infrastructure/jwt_service.go
@@ -1,7 +1,7 @@
 package Infrastructure
 
 import (
-	domain "AAiT-backend-group-8/Domain"
+	interfaces "AAiT-backend-group-8/Interfaces"
 	"errors"
 
 	"github.com/dgrijalva/jwt-go"
@@ -12,7 +12,7 @@ type TokenService struct {
 	SecretKey string
 }
 
-func NewTokenService(secretKey string) domain.ITokenService {
+func NewTokenService(secretKey string) interfaces.ITokenService {
 	return &TokenService{SecretKey: secretKey}
 }
 
@@ -60,7 +60,7 @@ func (ts *TokenService) GetClaimsOfToken(tokenString string) (map[string]interfa
 	if ok && token.Valid {
 		return claims, nil
 	}
-	// chnange claims to norma map
+	// change claims to norma map
 	myMap := make(map[string]interface{})
 	for key, value := range claims {
 		myMap[key] = value
