+++ backend/AAiT-backend-group-8/Interfaces/comment_interfaces.go
@@ -1,0 +1,2 @@
+package interfaces
+
