+++ backend/AAiT-backend-group-8/Interfaces/like_interfaces.go
@@ -1,0 +1,1 @@
+package interfaces
