+++ backend/AAiT-backend-group-8/Usecase/blog_usecase.go
@@ -2,16 +2,17 @@ package usecase
 
 import (
 	"AAiT-backend-group-8/Domain"
+	interfaces "AAiT-backend-group-8/Interfaces"
 	"time"
 
 	"go.mongodb.org/mongo-driver/bson/primitive"
 )
 
 type BlogUseCase struct {
-	repo Domain.IBlogRepository
+	repo interfaces.IBlogRepository
 }
 
-func NewBlogUseCase(repo Domain.IBlogRepository) *BlogUseCase {
+func NewBlogUseCase(repo interfaces.IBlogRepository) *BlogUseCase {
 	return &BlogUseCase{repo: repo}
 }
